Add 'before' support
TestCase::Spec can easily bleed state between tests because it has no
notion of cleaning up after each test.

Add the ability to define routines that must be run before and after each
test, following the behaivour of Test::Spec where possible.

'before' is now available to support both 'each' and 'all', but 'all'
has a caveat where it is run immediately meaning we have to be careful
that these are placed at the appropriate time in the test suite.