Add offset to Test::Builder::caller
TestCase::Spec is overriding part of Test::Builder so it can inject test
names into the TAP output. This has resulted in Test::Builder incorrectly
locating the line within a test file.

Add an offset to the caller level so that the test suite reports failed tests
within the test script rather than the test frameworks.